Refactoring. Temporarily remove app communication.
