+++ src/common.h
@@ -1,0 +1,12 @@
+/***** Useful Macros *****/
+	
+#define MAX(a, b) (( a > b)? a : b)
+	
+/***** Persist Keys *****/
+#define PERSIST_BOTTLE 1
+#define PERSIST_DIAPER 2
+#define PERSIST_MOON_START 3
+#define PERSIST_MOON_END 4
+
+
+
