+++ src/ui.h
@@ -1,0 +1,20 @@
+#include <pebble.h>
+#include <time.h>
+#include <math.h>
+#include <string.h>
+
+#include "common.h"
+
+void ui_diaper_change(time_t timestamp);
+
+void ui_feed(time_t timestamp);
+
+void ui_sleep_start(time_t timestamp);
+
+void ui_sleep_stop(time_t timestamp);
+
+void ui_init(ClickConfigProvider clickConfigProvider);
+
+void ui_update();
+
+void ui_deinit();
